IRdisplay-options {IRdisplay}	R Documentation

IRdisplay options

Description

Some options to control the formats display and prepare_mimebundle emit, and the function they use to display them.

Usage


irdisplay_option_defaults

Format

An object of class list of length 2.

Options

jupyter.display_mimetypes

The default is all MIME types supported by Jupyter.

jupyter.base_display_func

Function used by display and all display_<text> / display_<image> functions. Has the signature function(data, metadata = NULL). Per default emits a warning, and is set when running an IRkernel.

[Package IRdisplay version 0.7.0 Index]