display_<text> {IRdisplay}	R Documentation

Display a specific textual output

Description

Either data or file must be passed.

Usage


display_json(data = NULL, file = NULL)

display_javascript(data = NULL, file = NULL)

display_html(data = NULL, file = NULL)

display_markdown(data = NULL, file = NULL)

display_latex(data = NULL, file = NULL)

Arguments

data
The code or markup content as a character vector

file
The path to a file or a connection containing the content

See Also

display_<image>

Examples


## Not run: ## (Run inside of an IRkernel)
display_text('Just text')
display_markdown('[MD](http://commonmark.org) *formatted*')
display_javascript('execute(this)')
## End(Not run)


[Package IRdisplay version 0.7.0 Index]