publish_mimebundle {IRdisplay}	R Documentation

Display data by mimetype, with optional alternative representations.

Description

Calls the function stored as option value of jupyter.base_display_func. (see: IRdisplay-options)

Usage


publish_mimebundle(data, metadata = NULL)

Arguments

data
A named list mapping mimetypes to content (character or raw vectors)

metadata
A named list mapping mimetypes to named lists of metadata, e.g. list('image/png' = list(width = 5))

See Also

prepare_mimebundle

Examples


## Not run: ## (Run inside of an IRkernel)
publish_mimebundle(list('text/html' = '<h1>Hi!</h1>'))
publish_mimebundle(
  list('image/svg+xml' = '<svg xmlns="http://www.w3.org/2000/svg"><circle r="100"/></svg>'),
  list('image/svg+xml' = list(width = 100, height = 100)))
## End(Not run)


[Package IRdisplay version 0.7.0 Index]